requirements.txt
langchain
langchain-community
pypdf
python-dotenv
openai
chromadb 
tiktoken
langchain-openai